requirements.txt
google-generativeai
smolagents[openai]
chromadb
python-dotenv
ddgs
flask
flask-cors
elevenlabs
numpy
